feat(conversation): #WB-3881 add ellipsis
