+++ conversation/backend/src/main/resources/i18n/en.json
@@ -1,7 +1,7 @@
 {
     "add.attachment": "Ajouter une pièce jointe",
-    "add.signature" : "Add a signature",
-    "at": "à",
+    "add.signature": "Add a signature",
+    "at": "à:",
     "attachments": "Pièces jointes",
     "attachments.loading": "En cours de chargement",
     "cancel": "Cancel",
