+++ conversation/backend/src/main/resources/i18n/fr.json
@@ -1,7 +1,7 @@
 {
   "add.attachment": "Ajouter une pièce jointe",
   "add.signature": "Ajouter une signature",
-  "at": "à",
+  "at": "à :",
   "attachments": "Pièces jointes",
   "attachments.loading": "En cours de chargement",
   "author.avatar": "Avatar de l'auteur",
